Create Zi_Test5.csv
Quite a surprise! U.star is dominating!!

Well, maybe quite *not* a surprise, after all (if you have a glance to the Equations - as I should have done many times... :) ). But it's cozy, finding glimpse of Nature in the mathematical models used to describe it.

A nice question - a mere curiosity, not really a research idea - is, "why?" Why should the mixing height *estimate* respond so swiftly to relatively tiny changes in friction velocity?

And, is this somewhat indicative of the necessity to place the measurement point far from obstructions? *Much farther*, I mean, than suggested in WMO Report No.8?

And, what of *urban* measurements?? There I'd find difficult placing the sonic afar from obstructions, to the point the meaning of word "obstruction" is possibly questionable.

Last: finding an entire World in a small thing. A very Zen experience... :