Update pbl_wind.f90
Comment of SonicData % valid() changed to reflect treatment of H2O and CO2. Basically, the functional part of code has remained unchanged, because H2O and CO2 are allowed to be invalid without affecting sonic quadruple validity, as their acquisition process is too different - and includes a delay which blurs sense somewhat.